Devfreq: Added a Governor Whitelist
Many of the devfreq governors compiled in are unstable/incomplete, so we
don't want users switching to them. Compiling out the unused governors is
messy because many other drivers link to them, and there are not proper

Add a white list of the only governors intended for use to the governor
sysfs node in order to prevent users from enabling governors that unstable.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: MOVZX <movzx@yahoo.com>
Signed-off-by: rishabhrao <admin@rishabhrao.net>